chore: remove unnecessary scripts
